+++ src/ziggurat/kafka_consumer/consumer.clj
@@ -13,9 +13,9 @@
               ConsumerConfig/MAX_POLL_RECORDS_CONFIG         (int (or (:max-poll-records  consumer-group-config) 500))
               ConsumerConfig/SESSION_TIMEOUT_MS_CONFIG       (int (or (:session-timeout-ms-config  consumer-group-config) 60000))
               ConsumerConfig/ENABLE_AUTO_COMMIT_CONFIG       false
-              ConsumerConfig/KEY_DESERIALIZER_CLASS_CONFIG   (or (:key_deserializer_class_config consumer-group-config)
+              ConsumerConfig/KEY_DESERIALIZER_CLASS_CONFIG   (or (:key-deserializer-class-config consumer-group-config)
                                                                  "org.apache.kafka.common.serialization.ByteArrayDeserializer")
-              ConsumerConfig/VALUE_DESERIALIZER_CLASS_CONFIG (or (:value_deserializer_class_config consumer-group-config)
+              ConsumerConfig/VALUE_DESERIALIZER_CLASS_CONFIG (or (:value-deserializer-class-config consumer-group-config)
                                                                  "org.apache.kafka.common.serialization.ByteArrayDeserializer")})))
 
 (defn create-consumer